requirements.txt
flask
torch
torchvision
pillow
numpy
scikit-learn
matplotlib
seaborn
werkzeug
gunicorn
tqdm